# Removal of PixelPairStep for Phase 1 HI Tracking