--- a/RecoHI/HiTracking/python/HiTracking_cff.py
+++ b/RecoHI/HiTracking/python/HiTracking_cff.py
@@ -43,7 +43,6 @@ hiTracking_noRegitMu_wSplitting_Phase1 = cms.Sequence(
     *hiDetachedQuadStep#New iteration
     *hiDetachedTripletStep
     *hiLowPtTripletStep
-    *hiPixelPairStep #no CA seeding implemented
     )
 
 hiTracking = cms.Sequence(