After the inclusion of additional quadruplet iterations for Phase 1 tracking in PR #18646, we have decided to remove the pixel pair step from the phase 1 HI tracking workflow (similar to what was done for pp tracking).  Studies [here](https://indico.cern.ch/event/646309/contributions/2625076/attachments/1475706/2285895/PixelPairStudy.pdf) show that this iteration is only recovering 1-2% efficiency, while also contributing ~1% fake rate.  

Timing studies show a decrease in tracking time in 0-10% events from ~109 s to ~99 s, and from 10.3 to 9.7s in inclusive events.

Testing can be done using the HI Phase 1 relval proposed in PR #19089.

@mandrenguyen