@slava77 2017-06-19
@cmsbuild please test