@cmsbuild 2017-06-19
The tests are being triggered in jenkins.