@slava77 2017-06-19
Regarding

> we have decided to remove the pixel pair step from the phase 1 HI tracking workflow (similar to what was done for pp tracking).

you may want to check the latest developments in the pp reco in #19051 
pixel pair iteration is active in phase-1 pp reco