@cmsbuild 2017-06-19
+1
Tested at: 1e7da44f6d9d8a6a8ae2118aa4caf057ef1e0c6a
https://cmssdt.cern.ch/SDT/jenkins-artifacts/pull-request-integration/PR-19338/20737/summary.html