@slava77 2017-06-21
@cmsbuild please test

comparisons never showed up