@cmsbuild 2017-06-21
The tests are being triggered in jenkins.