@slava77 2017-06-21
@abaty 
please clarify if there is a follow up to https://github.com/cms-sw/cmssw/pull/19338#issuecomment-309437958

Are you planning to mitigate the inefficiency in a similar way as in pp (using pixelPair step)?