@cmsbuild 2017-06-21
Comparison is ready
https://cmssdt.cern.ch/SDT/jenkins-artifacts/pull-request-integration/PR-19338/20804/summary.html

Comparison Summary:
 - You potentially added 2 lines to the logs
 - Reco comparison results: 0 differences found in the comparisons
 - DQMHistoTests: Total files compared: 22
 - DQMHistoTests: Total histograms compared: 1803229
 - DQMHistoTests: Total failures: 15035
 - DQMHistoTests: Total nulls: 0
 - DQMHistoTests: Total successes: 1788028
 - DQMHistoTests: Total skipped: 166
 - DQMHistoTests: Total Missing objects: 0
 - Checked 90 log files, 14 edm output root files, 22 DQM output files