@abaty 2017-06-21
@slava77 

We are doing some quick investigations to see if this type of mitigation is worthwhile for us.  We should have an answer by the end of tomorrow.