@abaty 2017-06-23
@slava77 
After some review, we believe we would benefit from some similar mitigation strategy.  As this will involve keeping the pair step in some form, I will close this PR and make a new one once we have settled on an effective setup.